Console-master/requirements.txt
#apturl==0.5.2
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
bcrypt==3.2.0
blinker==1.4
#Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
#command-not-found==0.3
cryptography==3.4.8
cssselect2==0.6.0
#cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
#defer==1.0.6
distro==1.7.0
#distro-info===1.1build1
Django==4.1
duplicity==0.8.21
fasteners==0.14.1
fonttools==4.37.1
future==0.18.2
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
imageio==2.21.1
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.3
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
networkx==2.8.6
numpy==1.23.2
oauthlib==3.2.0
olefile==0.46
opencv-python==4.6.0.66
oscrypto==1.3.0
packaging==21.3
pandas==1.4.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyPDF3==1.0.6
pyRFC3339==1.1
pyserial==3.5
python-apt==2.3.0+ubuntu2.1
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
reportlab==3.6.8
requests==2.25.1
scikit-image==0.19.3
scipy==1.9.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
systemd-python==234
tifffile==2022.8.12
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.2
tzlocal==4.2
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritools==4.0.0
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xhtml2pdf==0.2.8
xkit==0.0.0
zipp==1.0.0
